ADD new forcefield to implement plastic behaviour
Adds a new forcefield inheriting from AdaptiveBeamForceFieldAndMass,
implementing a plastic mechanical behaviour (allowing to simulate
permanent deformation). The idea of the forcefield is to override only
forcefield interface functions, keeping the interpolation methods
identical to AdaptiveBeamFFAndMass.
In the same spirit, we will distinguish the spline interpolation, used
to handle the topology, and specific shape functions for the computation
of (tangent) stiffness.

In this commit, only the base skeleton of the component is added.