Add new interpolation component for plasticity
The idea is to keep all that is related to mechanics in the forcefield,
but use the interpolation method to retrieve the beams geometric
features (section type, dimensions, ...) which are necessary for
gaussian integration.
As beamInterpolation is already dense, we create a new component which
inherits from it, and implements only what is necessary to handle
plasticity.