Add classes and functions to handle gaussian integration
As plasticity implies that the material behaviour law becomes nonlinear
and history-dependant, the contribution of the forcefield to the global
mechanical system has to be recomputed at each time step, using Gaussian
integration (typically in addForce, and addKToMatrix, to get the current
internal forces/tangent stiffness matrix).

In this commit, we implement new methods in the forcefield to handle
Gaussian integration in a semi-generic way. This implementation relies
mostly on two internal classes:
  - Interval3, which describes the 3D intervals of integration for each
beam elements. These intervals are computed from the geometric data
available in Beam(Plastic)Interpolation. One Interval3 object is defined
and stored for each beam element in the forcefield.
  - GaussPoint3, which stores information regarding the Gauss points
used in each element for integration. At most, the plasticity
computations require to integrate 4th-degree polynomials in (x,y,z)
(with 2nd-degree crossed terms). We therefore choose to consider 3 gauss
points for each direction of integration, in each beam elements, which
represents in total 27 Gauss points per elements.
/!\ It is possible that fewer Gauss points are strictly necessary for
the integral to be exactly computed: this question should be addressed
in the future, to potentially reduce computational time /!\
At the moment, the GaussPoint3 structure only stores the point
coordinates and associated weights (one for each direction of
integration), but in future commits, they will be added
plasticity-related variables, such as plastic strain history,
mechanical state, and stress at last iteration.

Concretely, all the Gauss points are computed in init(). They are
initialised with basic coordinates and weights for integration on
[-1,1], then modified according to the actual integration intervals.
Finally, integration is actually performed in a method integrate(),
which applies a lambda function to each Gauss point of an element.

Pros of this implementation :
  - Plasticity is implemented entirely in the forcefield, only geometric
features are retrieved from the interpolation component.
  - The quadrature method can be used with any integral function, which
just has to be specified when required.

Cons :
  - The quadrature method is not generic, as it is defined inside of the
forcefield class.
  - It is probably redundant with other parts of the Sofa and/o plugins
code.