Change Interval3 and GaussPoint3 implementation
Both classes are made public, in order to be more easily tested in unit
tests. It seemed appropriate to keep them inside the forcefield for now,
but this could be modified in the future (for a more general use).
Corresponding methods to access their members are implemented.

All definitions are moved in the .inl file