Add computation of the tangent stiffness matrix
In plasticity, the tangent stiffness matrix plays the same role as the
usual stiffness matrix, but takes into account the nonlinear relation
between stress and strain, as well as the plasticity history.
NB: it is also possible to implement a so-called consistent tangent
operator (which is not the case here). In previous experiments, this
proved to make small difference in the computation, when other (bigger
?) approximations are otherwise used, but it should be further
investigated. An implementation of the consistent tangent operator is
available in the BeamPlastic Sofa plugin.