Remove beamtangentStiffness subclass
Instead, the tangent stiffness matrix is handled with the
localBeamMatrices class, inherited from AdaptiveBeamForceFieldAndMass.
This allows to spare the reimplementation of methods addDForce,
addKToMatrix, etc which would otherwise use tangent stiffness matrices
in the exact same way as elastic stiffness matrices.
It has to be clear however that the m_KXX matrices defined in
localBeamMatrices are not regular stiffness matrices but approximated
tangent stiffness matrices, in the case of the
AdaptivePlasticBeamForceField.