Refactor BeamPlasticInterpolation code
We finally delete the previously introduced plastic interpolation
component, and implement directly in BeamInterpolation the data
structure it was defining to export specific geometry properties of the
beam for plastic computation.
The problem with keeping a dedicated interpolation component is that the
force field parent class (AdaptiveBeamForceFieldAndMass) defines another
link with a more generic interpolation component (BeamInterpolation). In
the child class, all methods which are inherited as is and which use the
interpolation link look for the parent class link with
BeamInterpolation, which leads to execution errors.

A priori, this leaves two options:
  - Reimplementing explicitely all methods concerned by this issue
(addDForce, addKToMatrix, addMDx, ...), using the same code, which is
redundant.
  - Implement the data structure in BeamInterpolation and link the
AdaptivePlasticBeamForceField directly with BeamInterpolation. This is
the option chosen here.

NB: there might be another (better) solution, but I'm unfamiliar with how
Sofa links work, and can't figure it out at the moment.