Change the expression of the 'mechanical' shape functions
The previous expression of the shape function matrix and its derivative
were based on the assumption that the beam local frame was located at
the first beam node (node0). We correct these expressions, using
adequate boundary conditions, to obtain them expressed in a loal frame
located at the centre of the beam.