Add unit test to check forcefield methods
This commit introduces a unit test for AdaptivePlasticBeamForceField,
which is based on the methodes defined in ForceFieldTestCreation.h (in
SofaSimpleFem_test). Here, we focus on testing the coherence of the
addDForce method, by comparison with finite differences.
/!\ In order to work with the DoFs usually used with the beam model
(Rigid), line 245 in ForceFieldTestCreation.h has to be modified,
writing the dot product with the '*' operator rater than with the dot
method.